Refactor: consolidate decorators into new decorators folder, remove agenta_decorator & span files and integrate their contents into their individual decorator file
